+++ src/coreclr/scripts/cse_ml/jitml/__init__.py
@@ -2,3 +2,6 @@
 from .superpmi import SuperPmi
 from .jitenv import JitEnv
 from .machine_learning import JitRLModel
+from .observation import get_observation
+
+__all__ = ['SuperPmi', 'JitEnv', 'JitRLModel', 'get_observation']
